ServR/Kr_3/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
PyJWT>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
slowapi>=0.1.9
